moved snapshots
