+++ program/branches/init.go
@@ -107,27 +107,3 @@ func getDefaultBranch() string {
 func filterFn(branch string, text string) bool {
 	return strings.Contains(branch, strings.ToLower(text))
 }
-
-type snapshot struct {
-	remote bool
-	total  int
-
-	listNewSize        int
-	listTextInputValue string
-
-	width  int
-	height int
-}
-
-func (m Model) getSnapshot() snapshot {
-	return snapshot{
-		remote: m.remote,
-		total:  m.total,
-
-		listNewSize:        m.list.NewSize(),
-		listTextInputValue: m.list.TextInput.Value(),
-
-		width:  m.width,
-		height: m.height,
-	}
-}
