+++ program/commits/init.go
@@ -99,25 +99,3 @@ func getWidth(width int) int {
 func getHeight(height int) int {
 	return height - 2
 }
-
-type snapshot struct {
-	total int
-
-	listNewSize        int
-	listTextInputValue string
-
-	width  int
-	height int
-}
-
-func (m Model) getSnapshot() snapshot {
-	return snapshot{
-		total: m.total,
-
-		listNewSize:        m.list.NewSize(),
-		listTextInputValue: m.list.TextInput.Value(),
-
-		width:  m.width,
-		height: m.height,
-	}
-}
