+++ program/files/init.go
@@ -136,26 +136,3 @@ func getPath(fileStr string) string {
 func filterFn(path string, text string) bool {
 	return strings.Contains(path, strings.ToLower(text))
 }
-
-type snapshot struct {
-	total int
-
-	listNewSize        int
-	listTextInputValue string
-
-	width  int
-	height int
-}
-
-func (m Model) getSnapshot() snapshot {
-	width, height := m.getFilesAxis()
-	return snapshot{
-		total: m.total,
-
-		listNewSize:        m.list.NewSize(),
-		listTextInputValue: m.list.TextInput.Value(),
-
-		width:  width,
-		height: height,
-	}
-}
